Refactor task result tests and add serializer test
